feat: deleting category with budget control

+++ finance-service/src/main/java/com/smartwallet/financeservice/service/CategoryService.java
@@ -1,8 +1,10 @@
 package com.smartwallet.financeservice.service;
 
+import com.smartwallet.financeservice.client.BudgetClient;
 import com.smartwallet.financeservice.dto.request.CreateCategoryRequest;
 import com.smartwallet.financeservice.dto.response.CategoryResponse;
 import com.smartwallet.financeservice.entity.Category;
+import com.smartwallet.financeservice.exception.BudgetFoundForCategoryException;
 import com.smartwallet.financeservice.exception.CategoryAlreadyExistsException;
 import com.smartwallet.financeservice.exception.CategoryCannotDeleteBecauseTransactions;
 import com.smartwallet.financeservice.exception.CategoryNotFoundException;
@@ -22,6 +24,7 @@ public class CategoryService {
     private final CategoryRepository categoryRepository;
     private final CategoryMapper categoryMapper;
     private final FinancialTransactionRepository financialTransactionRepository;
+    private final BudgetClient budgetClient;
 
     @Transactional
     public CategoryResponse createCategory(
@@ -80,15 +83,33 @@ public class CategoryService {
     }
 
     @Transactional
-    public void deleteCategory(Long userId, Long categoryId){
+    public void deleteCategory(Long userId, Long categoryId, String accessToken){
         Category category = categoryRepository
                 .findByIdAndUserId(categoryId, userId)
                 .orElseThrow(
                         () -> new CategoryNotFoundException(categoryId)
                 );
-        boolean transactionExists = financialTransactionRepository.existsByUserIdAndCategory(userId,category);
+
+        boolean transactionExists =
+                financialTransactionRepository
+                        .existsByUserIdAndCategory(
+                                userId,
+                                category
+                        );
+
         if(transactionExists){
             throw new CategoryCannotDeleteBecauseTransactions();
+        }
+
+        Boolean isBudgetExists =
+                budgetClient
+                        .getCategoryBudget(
+                                categoryId,
+                                accessToken
+                        );
+
+        if(isBudgetExists){
+            throw new BudgetFoundForCategoryException();
         }
 
         categoryRepository.delete(category);
